# Add copy operator to fix memory benchmarks