--- a/napari/utils/colormaps/colormap.py
+++ b/napari/utils/colormaps/colormap.py
@@ -466,9 +466,11 @@ except ModuleNotFoundError:
     _zero_preserving_modulo = _zero_preserving_modulo_numpy
 else:
 
-    @numba.njit(parallel=True)
     def _zero_preserving_modulo(
-        values: np.ndarray, n: int, dtype: np.dtype, to_zero: int = 0
+        values: np.ndarray,
+        n: int,
+        dtype: np.dtype,
+        to_zero: int = 0,
     ) -> np.ndarray:
         """``(values - 1) % n + 1``, but with one specific value mapped to 0.
 
@@ -494,14 +496,51 @@ else:
             everywhere else.
         """
         result = np.empty_like(values, dtype=dtype)
+        _zero_preserving_modulo_impl(
+            values,
+            n,
+            result,
+            to_zero,
+        )
+        return result
 
+    @numba.njit(parallel=True)
+    def _zero_preserving_modulo_impl(
+        values: np.ndarray,
+        n: int,
+        out: np.ndarray,
+        to_zero: int = 0,
+    ) -> np.ndarray:
+        """``(values - 1) % n + 1``, but with one specific value mapped to 0.
+
+        This ensures (1) an output value in [0, n] (inclusive), and (2) that
+        no nonzero values in the input are zero in the output, other than the
+        ``to_zero`` value.
+
+        Parameters
+        ----------
+        values : np.ndarray
+            The dividend of the modulo operator.
+        n : int
+            The divisor.
+        out : np.ndarray
+            output array
+        to_zero : int, optional
+            A specific value to map to 0. (By default, 0 itself.)
+
+        Returns
+        -------
+        np.ndarray
+            The out: 0 for the ``to_zero`` value, ``values % n + 1``
+            everywhere else.
+        """
         for i in numba.prange(values.size):
             if values.flat[i] == to_zero:
-                result.flat[i] = 0
+                out.flat[i] = 0
             else:
-                result.flat[i] = (values.flat[i] - 1) % n + 1
+                out.flat[i] = (values.flat[i] - 1) % n + 1
 
-        return result
+        return out
 
 
 def minimum_dtype_for_labels(num_colors: int) -> np.dtype: