# References and relevant issues
closes #6529

# Description

This PR provides two changes to fix benchmarking of memory usage by layers. 

I have found that `asv` uses following code to calculate memory usage for an object. 

```python
asizeof([obj, copy(obj)]) - asizeof([obj, obj])
```

This approach allows not to count of global (or class) object into size of single object. But if the copy of an object is shallow, then the reported size is small. 

So in this PR, I define `__copy__` method of Base layer, that uses `as_layer_data_tuple` method and `create` class method to create a copy. It also creates copy of data, as it is a significant part of layer size. 

I also found a second problem. 

The `aszieof` function on output of jited version of `_zero_preserving_modulo` reports small values (like 128 bytes). But when I lift array allocation outside implementation, the memory is reported properly. So I have updated the `_zero_preserving_modulo` implementation to allocate memory before entering the jited code. 

Current main:

```
[50.00%] ··· benchmark_labels_layer.Labels2DSuite.mem_data                                                         ok
[50.00%] ··· ====== ============= =============
             --                dtype
             ------ ---------------------------
               n     numpy.uint8   numpy.int32
             ====== ============= =============
               16        384          1.15k
               32       1.15k         4.22k
               64       4.22k         16.5k
              128       16.5k         65.7k
              256       65.7k          262k
              512        262k         1.05M
              1024      1.05M         4.19M
              2048      4.19M         16.8M
              4096      16.8M         67.1M
             ====== ============= =============

[100.00%] ··· benchmark_labels_layer.Labels2DSuite.mem_layer                                                        ok
[100.00%] ··· ====== ============= =============
              --                dtype
              ------ ---------------------------
                n     numpy.uint8   numpy.int32
              ====== ============= =============
                16       3.29k         3.29k
                32       3.29k         3.29k
                64       3.29k         3.29k
               128       3.29k         3.29k
               256       3.29k         3.29k
               512       3.29k         3.29k
               1024      3.29k         3.29k
               2048      3.29k         3.29k
               4096      3.29k         3.29k
              ====== ============= =============
```

with added copy:
```
[50.00%] ··· benchmark_labels_layer.Labels2DSuite.mem_data                                                         ok
[50.00%] ··· ====== ============= =============
             --                dtype
             ------ ---------------------------
               n     numpy.uint8   numpy.int32
             ====== ============= =============
               16        384          1.15k
               32       1.15k         4.22k
               64       4.22k         16.5k
              128       16.5k         65.7k
              256       65.7k          262k
              512        262k         1.05M
              1024      1.05M         4.19M
              2048      4.19M         16.8M
              4096      16.8M         67.1M
             ====== ============= =============

[100.00%] ··· benchmark_labels_layer.Labels2DSuite.mem_layer                                                        ok
[100.00%] ··· ====== ============= =============
              --                dtype
              ------ ---------------------------
                n     numpy.uint8   numpy.int32
              ====== ============= =============
                16        188k          189k
                32        189k          192k
                64        192k          205k
               128        205k          254k
               256        254k          450k
               512        450k         1.24M
               1024      1.24M         4.38M
               2048      4.38M          17M
               4096       17M          67.3M
              ====== ============= =============
```

with copy and refactored `_zero_preserving_modulo`
```
[50.00%] ··· benchmark_labels_layer.Labels2DSuite.mem_data                                                         ok
[50.00%] ··· ====== ============= =============
             --                dtype
             ------ ---------------------------
               n     numpy.uint8   numpy.int32
             ====== ============= =============
               16        384          1.15k
               32       1.15k         4.22k
               64       4.22k         16.5k
              128       16.5k         65.7k
              256       65.7k          262k
              512        262k         1.05M
              1024      1.05M         4.19M
              2048      4.19M         16.8M
              4096      16.8M         67.1M
             ====== ============= =============

[100.00%] ··· benchmark_labels_layer.Labels2DSuite.mem_layer                                                        ok
[100.00%] ··· ====== ============= =============
              --                dtype
              ------ ---------------------------
                n     numpy.uint8   numpy.int32
              ====== ============= =============
                16        188k          189k
                32        189k          193k
                64        192k          209k
               128        205k          270k
               256        254k          516k
               512        450k          1.5M
               1024      1.24M         5.43M
               2048      4.38M         21.2M
               4096       17M          84.1M
              ====== ============= =============
```