@brisvag 2023-12-12
Shouldn't this be the `deepcopy` behaviour then?