@jni 2023-12-12
I'm not super concerned about exactly following the copy/deepcopy semantics. This is mostly for internal usage I think...