@jni 2023-12-12
(But if you have specific suggestions @brisvag that are easy to implement here I of course wouldn't be opposed to including them!)