@brisvag 2023-12-12
If it's just for internal use which is deviating from the standard of copy/deepcopy, then I would suggest making a private method for it, like `_copy_with_data`... as a user I would expect `copy(layer)` and `deepcopy(layer)` to follow the nodata/data intuition.