@Czaki 2023-12-12
I have found that `asizeof` wrongly calculated the size of array allocated in jited function